feat(LineLimitScene): refactor error message to work for all tabs
